backend/app/ai/embedder.py
from sentence_transformers import SentenceTransformer
import numpy as np

# This model runs locally — no API key needed.
# 'all-MiniLM-L6-v2' is fast, small, and good enough for industrial text.
# Produces 384-dimensional vectors.

_model = None

def get_model() -> SentenceTransformer:
    global _model
    if _model is None:
        print("Loading embedding model (first time only)...")
        _model = SentenceTransformer("all-MiniLM-L6-v2")
        print("Embedding model loaded.")
    return _model

def embed_text(text: str) -> list[float]:
    model = get_model()
    vector = model.encode(text, normalize_embeddings=True)
    return vector.tolist()

def embed_batch(texts: list[str]) -> list[list[float]]:
    model = get_model()
    vectors = model.encode(texts, normalize_embeddings=True, batch_size=32)
    return vectors.tolist()