backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.35.0
neo4j==5.28.0
python-dotenv==1.0.1
groq==0.15.0
sentence-transformers==3.4.1
opensearch-py==2.8.0
psycopg[binary]==3.3.3
pydantic==2.11.0
httpx==0.27.0